Readded expiration on routes, moved files around, and cleanup
