+++ asabr/src/distance/sabr.rs
@@ -43,7 +43,7 @@ impl<NM: NodeManager, CM: ContactManager> Distance<NM, CM> for SABR {
         _bundle: &Bundle,
     ) -> Ordering {
         super::cmp_by(first, second, |path| {
-            (path.arrival_time.end, path.hop_count)
+            (path.arrival_time.end, path.hop_count, path.expiration)
         })
         // TODO: Readd expiration
     }
